+++ client-desktop/src/main/java/main/org/jugendhackt/HAWebDesktop/Gui/ChatGui.java
@@ -1,30 +1,39 @@
-package org.jugendhackt.HAWebDesktop.Gui;
+package main.org.jugendhackt.HAWebDesktop.Gui;
 
-import org.jugendhackt.HAWebDesktop.Util;
+import main.HAWebDesktop.Gui.Gui;
+import main.HAWebDesktop.Util;
 
 import javax.swing.*;
 import java.awt.*;
 
-public class KNOWHOWGui {
-    private JPanel KnowhowPanel;
+public class ChatGui {
+    private JPanel Chatpane;
     private JButton knowHowButton;
     private JButton homeworkButton;
     private JButton chatButton;
     private JButton homeButton;
     private JPanel topbar;
+    private JList Chatcontent;
+    private JTextField messfield;
+    private JButton submitMessage;
+    private JButton knowhowButton;
 
-    public JPanel getKnowhowPanel() {
+
+    public JPanel getChatpane() {
         homeButton = Util.makeTopBarClick(homeButton, Gui.HOME);
         knowHowButton = Util.makeTopBarClick(knowHowButton, Gui.KNOWHOW);
         homeworkButton = Util.makeTopBarClick(homeworkButton, Gui.HOMEWORK);
         chatButton = Util.makeTopBarClick(chatButton, Gui.CHAT);
 
         homeButton = makebtn(homeButton, true);
-        knowHowButton = makebtn(knowHowButton, false);
+        knowHowButton = makebtn(knowHowButton, true);
         homeworkButton = makebtn(homeworkButton, true);
-        chatButton = makebtn(chatButton, true);
+        chatButton = makebtn(chatButton, false);
+        DefaultListModel model = new DefaultListModel();
+        model.addElement("Nachricht!");
+        Chatcontent.setModel(model);
 
-        return KnowhowPanel;
+        return Chatpane;
     }
 
     public JButton makebtn(JButton button, boolean back) {
@@ -39,4 +48,6 @@ public class KNOWHOWGui {
 
         return button;
     }
+
+
 }
