+++ client-desktop/src/main/java/main/org/jugendhackt/HAWebDesktop/Gui/Gui.java
@@ -1,4 +1,4 @@
-package org.jugendhackt.HAWebDesktop.Gui;
+package main.HAWebDesktop.Gui;
 
 import javax.swing.*;
